feat: expose patchActor()
`patchActor` is a utility to mutate an existing `ActorRef` allowing addition of an inspector and/or new logger.

Inspectors are additive, so adding a new inspector is not destructive.

The logger needs to be replaced with a new function that calls the original logger and the new logger. If the `ActorRef` is a the root actor, the logger will be set at the system level; otherwise it will only be set for the particular `ActorRef`. This diverges from XState's behavior, which currently intends to only support a system-level logger.

The new `unpatchActor()` is an experimental API available to undo the changes made in `patchActor()`.

fix(logging): fix destructive logging

**xstate-audition** will no longer destroy an Actor's existing logger; it will now compose loggers correctly.